requirements.txt
# === 기반 패키지 (숫자 연산/데이터핸들링) ===
numpy>=1.26,<3
pandas>=2.2.3,<2.3
six>=1.16,<2

# === 파일 처리 & IO ===
openpyxl>=3.1.2,<3.2
fastparquet>=2024.5.0,<2025
olefile>=0.47,<0.48
python-docx>=1.1.0,<1.3
python-pptx>=1.0.0,<1.1
pdfminer.six>=20231228,<20240600
PyMuPDF>=1.26.4,<1.27
pdfplumber>=0.11.4,<0.12
python-dotenv>=1.0.1,<1.1

# === 번역/텍스트 유틸 ===
deep-translator>=1.11.4,<2
rank-bm25>=0.2.2,<0.3

# === ML & AI 모델 ===
joblib>=1.4.2,<1.5
scikit-learn>=1.6.0,<1.7
sentence-transformers>=3.0.1,<4
faiss-cpu>=1.9,<2   # (Windows 3.13 → 실제 설치는 faiss-cpu==1.9.0.post1 권장)
faster-whisper>=1.0,<1.1   # 회의 비서 STT 백엔드
transformers>=4.44,<5   # KoBART summarisation
# pykospacing는 macOS/arm64 wheel 미제공 -> 필요시 수동설치
kosspacing>=0.5,<0.6; python_version >= "3.12"   # torch 기반 띄어쓰기 대체
# py-hanspell은 구버전 pip API 의존 (필요 시 수동 설치)

# === 시스템 유틸 ===
watchdog>=4.0.0,<4.1
pywin32; platform_system == "Windows"
# pyhwp는 Python3 최신 버전 미지원 → 필요 시 별도 설치

# === 기타 ===
tqdm>=4.66,<5
pytest>=8.0,<9
jsonschema>=4.18.0,<5
customtkinter>=5.2,<6
fastapi>=0.117,<0.118   # 테스트용 API 서버 의존성
httpx>=0.28,<0.29       # FastAPI TestClient 의존성
